sec-performance/requirements.txt
paho-mqtt
PyJWT
numpy
cryptography
certifi